refactor(translator): drop the `.deprecated.ts` role tag — it cannot mark most of what it promised
The tag was added two commits ago so a directory listing would show what dies at
the next major without opening the ledger. Measured on this branch, it does not:
the ledger names five files in the provider layer and exactly one can carry the
tag.

The reason is structural, not an oversight. A dotted suffix has one slot, and
"deprecated" is a lifecycle attribute orthogonal to role, so the two compete and
role always wins. `OpenAITranslation.provider.ts` exports nothing that survives
the next major and cannot be tagged, because it is a `.provider.ts`.
`loadOpenAIClient.ts` and `runDryRun.ts` die too, and the convention explicitly
exempts topic-named helpers from tags at all. What is left able to carry the tag
is whatever has no other role — by definition the least significant file in the
set, here 27 lines of pure delegation.

So the marking is not merely incomplete, it is incomplete *systematically* and in
the wrong direction: a reader opening `openai/` sees one file marked out of three
that die, and concludes the other two stay.

Lifecycle stays where it already was, and where it is machine-readable: a
`@deprecated` tag on every deprecated symbol, `Code refs` in the ledger, and the
`// Deprecated exports` block in `src/index.ts`. If a listing-level signal is ever
wanted, the level that can hold it is a directory, not a suffix.

The file becomes `OpenAITranslationLegacy.provider.ts`, keeping the split that
made the removal a file deletion. One imperfection, stated rather than hidden: the
name does not compose to the symbol it exports the way the convention's other
files do, because the name that would compose is taken by the factory file.

The task doc records the reversal and strikes the risk assessment that said this
tag would pay off across all five deprecations. It paid off across one.

No behaviour change. 1291 tests across 107 files, check-types clean, lint 0/0,
all 13 ledger code refs resolve.

Refs #99
