+++ packages/payload-plugin-translator/src/translation-providers/openai/OpenAITranslationLegacy.provider.ts
